Present sila amounts as integers to end user. If any decimal is present, then it is truncated from it.

+++ SilaAPI/silamoney/client/api/SilaApi.cs
@@ -119,7 +119,7 @@ namespace SilaAPI.silamoney.client.api
         /// <param name="businessUuid">Optional. UUID of a business with an approved ACH name.</param>
         /// <param name="processingType">Optional.</param>
         /// <returns>ApiResponse&lt;object&gt; object with the server response</returns>
-        public ApiResponse<object> IssueSila(string userHandle, float amount, string userPrivateKey, string accountName = "default", string descriptor = null, string businessUuid = null, ProcessingType? processingType = null)
+        public ApiResponse<object> IssueSila(string userHandle, int amount, string userPrivateKey, string accountName = "default", string descriptor = null, string businessUuid = null, ProcessingType? processingType = null)
         {
             BankTransactionMessage body = new BankTransactionMessage(userHandle, amount, this.Configuration.AppHandle, accountName, descriptor, businessUuid, processingType, BaseMessage.Message.IssueMsg);
             var path = "/issue_sila";
@@ -176,7 +176,7 @@ namespace SilaAPI.silamoney.client.api
         /// <param name="businessUuid"></param>
         /// <param name="processingType"></param>
         /// <returns>ApiResponse&lt;object&gt; object with the server response</returns>
-        public ApiResponse<object> RedeemSila(string userHandle, float amount, string userPrivateKey, string accountName = "default",
+        public ApiResponse<object> RedeemSila(string userHandle, int amount, string userPrivateKey, string accountName = "default",
             string descriptor = null, string businessUuid = null, ProcessingType? processingType = null)
         {
             BankTransactionMessage body = new BankTransactionMessage(userHandle, amount, Configuration.AppHandle, accountName, descriptor, businessUuid, processingType, BaseMessage.Message.RedeemMsg);
@@ -302,7 +302,7 @@ namespace SilaAPI.silamoney.client.api
         /// <param name="descriptor"></param>
         /// <param name="businessUuid"></param>
         /// <returns>ApiResponse&lt;object&gt; object with the server response</returns>
-        public ApiResponse<object> TransferSila(string userHandle, float amount, string destinationHandle, string userPrivateKey, string destinationAddress = null, string destinationWallet = null,
+        public ApiResponse<object> TransferSila(string userHandle, int amount, string destinationHandle, string userPrivateKey, string destinationAddress = null, string destinationWallet = null,
             string descriptor = null, string businessUuid = null)
         {
             TransferMsg body = new TransferMsg(userHandle, amount, destinationHandle, Configuration.AppHandle, destinationAddress, destinationWallet, descriptor, businessUuid);
